firmware-tools/ptgen: do not truncate the output file
At the moment, the image file operated on is always truncated
to 512 bytes. This limits the usage of ptgen somewhat.

This change allows to use ptgen on existing, larger image files.
In this case, only the partition table is written/updated while
the rest of the image file is left alone.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>