firmware-tools/ptgen: allow individual partition alignment
At the moment, the alignment parameter can only be given once and affects
all partitions the same way. For example, it's not possible to align the
first partition to 1 MiB start offet, while aligning the next ones to
a 4 MiB boundary.

This can be useful for example when defining a small bootloader partition
within the first erase block on an eMMC, but then aligning the rootfs to
next start of erase block.

This change splits the global alignment parameter to individual ones
which thus allows individual partition alignment.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>